% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beds_per_parcel}
\alias{beds_per_parcel}
\title{Bed counts per parcel in King County}
\format{
A data.table
}
\usage{
beds_per_parcel
}
\description{
A dataset containing information related to all KC parcels that seem to have
at least one bed in/on them (and studios). Provides information about beds in
apartment buildings, condos, and residential (e.g. <= 4 units) on a given parcel.
It also contains the GEOID for the 2020 census tract that a given parcel
geographically overlaps with the most (via st_join(largest = TRUE))
}
\keyword{datasets}
