% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{housing_parcels_shp}
\alias{housing_parcels_shp}
\title{Geometries of parcels with beds in King County}
\format{
A sf data frame
}
\usage{
housing_parcels_shp
}
\description{
A dataset containing information related to all KC parcels that seem to have
at least one bed in/on them (and studios)
}
\keyword{datasets}
