% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parcel_pop}
\alias{parcel_pop}
\title{Parcel level population estimates}
\format{
A sf (points) data frame
}
\usage{
parcel_pop
}
\description{
Parcel level population estimates based on the following model:
tract_pop ~ 0 + apt_beds + condo_beds + residential_beds
}
\details{
The model is fit as a simple lm (poisson models were not downscaling as well)
at the tract level and downscaled via the coefficients to the parcel level.

The initial parcel level estimates are scaled to recover tract level estimates.
}
\keyword{datasets}
