% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcoords}
\alias{pcoords}
\title{Centriods of parcels with beds in King County}
\format{
A sf data frame
}
\usage{
pcoords
}
\description{
A dataset containing information related to all KC parcels that seem to have
at least one bed in/on them (and studios)
}
\keyword{datasets}
