% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_pop.R
\name{point_pop}
\alias{point_pop}
\title{Fetch population points}
\usage{
point_pop(
  year,
  geog_window,
  table = DBI::Id(schema = "ref", table = "kc_point_pop"),
  con = apde.data::authenticate_hhsaw()
)
}
\arguments{
\item{year}{numeric vector of the year(s) to pull. Defaults to most recent.}

\item{geog_window}{A sf data.frame (presumably of polygons). Only points intersecting the union of the input shapes will be returned.}

\item{table}{DBI::Id of the table containing the point pop}

\item{con}{a DBI data base connection}
}
\value{
An sf data.frame of points and a column describing the number of people approximately at that point
}
\description{
Fetch population points
}
